% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.r
\name{chart_bp}
\alias{chart_bp}
\title{Control chart to monitor sample mean of beta prime observations}
\usage{
chart_bp(data, alpha = 0.0027, mu = NULL, phi = NULL)
}
\arguments{
\item{data}{The data points (matrix)}

\item{alpha}{The significance level for control limits}

\item{mu}{The mean of the distribution}

\item{phi}{The shape parameter}
}
\value{
An ggplot2 object with the control limits plot
}
\description{
Control chart to monitor sample mean of beta prime observations
}
\examples{
# data <- r_bp(500, 5, 1, 1.7)
# chart_bp(data, 0.0027, 1, 1.7)

}
