% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.r
\name{chart_g}
\alias{chart_g}
\title{Control chart to monitor sample mean of gamma observations}
\usage{
chart_g(data, alpha = 0.0027, mu = NULL, k = NULL)
}
\arguments{
\item{data}{The data points (matrix)}

\item{alpha}{The significance level for control limits}

\item{mu}{The mean of the distribution}

\item{k}{The shape parameter}
}
\value{
An ggplot2 object with the control limits plot
}
\description{
Control chart to monitor sample mean of gamma observations
}
\examples{
# data <- r_g(500, 5, 1, 1.7)
# chart_g(data, 0.0027, 1, 1.7)

}
