% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversa_gaussiana.r
\name{chart_ig}
\alias{chart_ig}
\title{Control chart to monitor sample mean of Inverse Gaussian observations}
\usage{
chart_ig(data, alpha = 0.0027, mu = NULL, lambda = NULL)
}
\arguments{
\item{data}{The data points (matrix)}

\item{alpha}{The significance level for control limits}

\item{mu}{The mean of the distribution}

\item{lambda}{The shape parameter}
}
\value{
An ggplot2 object with the control limits plot
}
\description{
Control chart to monitor sample mean of Inverse Gaussian observations
}
\examples{
# data <- r_ig(500, 5, 1, 1.7)
# chart_ig(data, 0.0027, 1, 1.7)

}
