% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{chart_n}
\alias{chart_n}
\title{Control chart to monitor sample mean of normal observations}
\usage{
chart_n(data, alpha = 0.0027, mu = NULL, sd = NULL)
}
\arguments{
\item{data}{The data points (matrix)}

\item{alpha}{The significance level for control limits}

\item{mu}{The mean of the distribution}

\item{sd}{The standard deviation of the distribution}
}
\value{
An ggplot2 object with the control limits plot
}
\description{
Control chart to monitor sample mean of normal observations
}
\examples{
# data <- r_n(500, 5, 1, 1.7)
# chart_n(data, 0.0027, 1, 1.7)

}
