% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.r
\name{d_sample_mean_bp}
\alias{d_sample_mean_bp}
\title{Calculate the sample mean of a Beta Prime distribution}
\usage{
d_sample_mean_bp(x, n, mu, phi)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{n}{The number of observations in each sample}

\item{mu}{The mean parameter of the Beta Prime distribution}

\item{phi}{The shape parameter of the Beta Prime distribution}
}
\value{
The sample mean of the Beta Prime distribution
}
\description{
Calculate the sample mean of a Beta Prime distribution
}
\examples{
x <- c(0.2, 0.3, 0.4)
d_sample_mean_bp(x, n = 3, mu = 1, phi = 2)

}
