% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.r
\name{d_sample_mean_g}
\alias{d_sample_mean_g}
\title{Calculate the gamma distribution for a given sample mean}
\usage{
d_sample_mean_g(x, n, mu, k)
}
\arguments{
\item{x}{The sample mean}

\item{n}{The sample size}

\item{mu}{The mean of the gamma distribution}

\item{k}{The shape parameter of the gamma distribution}
}
\value{
The probability density function value
}
\description{
This function calculates the probability density function of the gamma distribution
for a given sample mean.
}
\examples{
d_sample_mean_g(x = 2, n = 10, mu = 5, k = 2)

}
