% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversa_gaussiana.r
\name{d_sample_mean_ig}
\alias{d_sample_mean_ig}
\title{Calculate the inverse Gaussian probability density function}
\usage{
d_sample_mean_ig(x, n, mu, lambda)
}
\arguments{
\item{x}{The value at which to evaluate the PDF.}

\item{n}{The number of observations.}

\item{mu}{The mean parameter of the distribution.}

\item{lambda}{The shape parameter of the distribution.}
}
\value{
The PDF value at the given x.
}
\description{
This function calculates the probability density function (PDF) of the inverse Gaussian distribution.
}
\examples{
d_sample_mean_ig(x = 2, n = 10, mu = 1, lambda = 2)

}
