% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{d_sample_mean_n}
\alias{d_sample_mean_n}
\title{Calculate the density of the sample mean for a normal distribution}
\usage{
d_sample_mean_n(x, n, mu, sd)
}
\arguments{
\item{x}{The value at which to evaluate the density}

\item{n}{The sample size}

\item{mu}{The mean of the distribution}

\item{sd}{The standard deviation of the distribution
Construct control limits using echarts4r}
}
\value{
The density of the sample mean
}
\description{
Calculate the density of the sample mean for a normal distribution
}
\examples{
d_sample_mean_n(x = 2, n = 10, mu = 0, sd = 1)

}
