% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.r
\name{r_bp}
\alias{r_bp}
\title{Generate random samples from a Beta Prime distribution}
\usage{
r_bp(lots, n, mu, phi, ...)
}
\arguments{
\item{lots}{The number of samples to generate}

\item{n}{The number of observations in each sample}

\item{mu}{The mean parameter of the Beta Prime distribution}

\item{phi}{The shape parameter of the Beta Prime distribution}

\item{...}{Additional arguments to be passed to extraDistr::rbetapr}
}
\value{
A matrix of random samples from the Beta Prime distribution
}
\description{
Generate random samples from a Beta Prime distribution
}
\examples{
r_bp(lots = 10, n = 5, mu = 1, phi = 2)

}
