% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.r
\name{r_g}
\alias{r_g}
\title{Generate random samples from the gamma distribution}
\usage{
r_g(lots, n, mu, k, ...)
}
\arguments{
\item{lots}{The number of lots}

\item{n}{The sample size}

\item{mu}{The mean of the gamma distribution}

\item{k}{The shape parameter of the gamma distribution}

\item{...}{Additional arguments to be passed to \code{rgamma()} from the \code{stats} package}
}
\value{
A matrix of random samples
}
\description{
This function generates random samples from the gamma distribution.
}
\examples{
r_g(lots = 5, n = 10, mu = 5, k = 2)
r_g(lots = 10, n = 20, mu = 10, k = 3)

}
