% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversa_gaussiana.r
\name{r_ig}
\alias{r_ig}
\title{Generate random samples from the inverse Gaussian distribution}
\usage{
r_ig(lots, n, mu, lambda)
}
\arguments{
\item{lots}{The number of lots to generate.}

\item{n}{The number of observations per lot.}

\item{mu}{The mean parameter of the distribution.}

\item{lambda}{The shape parameter of the distribution.}
}
\value{
A matrix of random samples from the inverse Gaussian distribution.
}
\description{
This function generates random samples from the inverse Gaussian distribution.
}
\examples{
r_ig(lots = 5, n = 10, mu = 1, lambda = 2)

}
