% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{r_n}
\alias{r_n}
\title{Generate random samples from a normal distribution}
\usage{
r_n(lots, n, mu, sd)
}
\arguments{
\item{lots}{The number of lots to generate}

\item{n}{The sample size}

\item{mu}{The mean of the distribution}

\item{sd}{The standard deviation of the distribution}
}
\value{
A matrix of random samples
}
\description{
Generate random samples from a normal distribution
}
\examples{
r_n(lots = 100, n = 10, mu = 0, sd = 1)

}
