% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prime.r
\name{stats_bp}
\alias{stats_bp}
\title{Calculate statistical measures for a Beta Prime control chart}
\usage{
stats_bp(data, alpha = 0.0027, mu = NULL, phi = NULL, ...)
}
\arguments{
\item{data}{A matrix of observed data}

\item{alpha}{The significance level for the control limits}

\item{mu}{The mean parameter of the Beta Prime distribution (optional)}

\item{phi}{The shape parameter of the Beta Prime distribution (optional)}

\item{...}{Additional arguments to be passed to limits_bp}
}
\value{
A list containing the estimated parameters, control limits, and statistical measures
}
\description{
Calculate statistical measures for a Beta Prime control chart
}
\examples{
set.seed(0)
x <- r_bp(lots = 1000, n = 10, mu = 1, phi = 1.7)

# Estimate for maximum likelihood
stats_bp(data = x, alpha = 0.0027)

# Useing the true parameters
stats_bp(data = x, alpha = 0.0027, mu = 1, phi = 1.7)

}
