% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.r
\name{stats_g}
\alias{stats_g}
\title{Calculate statistical measures for the gamma distribution}
\usage{
stats_g(data, alpha = 0.0027, mu = NULL, k = NULL, ...)
}
\arguments{
\item{data}{The data matrix}

\item{alpha}{The significance level}

\item{mu}{The mean of the gamma distribution (optional)}

\item{k}{The shape parameter of the gamma distribution (optional)}

\item{...}{Additional arguments to be passed \code{lbfgs::lbfgs()}}
}
\value{
A list containing the statistical measures and other information
}
\description{
This function calculates statistical measures for the gamma distribution
based on the given data.
}
\examples{

set.seed(0)
x <- r_g(lots = 1000, n = 10, mu = 1, k = 1.7)

# Estimate for maximum likelihood
stats_g(data = x)

# Using the true parameters
stats_g(data = x, mu = 1, k = 1.7)

}
