% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inversa_gaussiana.r
\name{stats_ig}
\alias{stats_ig}
\title{Calculate statistical measures for the inverse Gaussian control chart}
\usage{
stats_ig(data, alpha = 0.0027, mu = NULL, lambda = NULL, ...)
}
\arguments{
\item{data}{The data matrix.}

\item{alpha}{The significance level.}

\item{mu}{The mean parameter of the distribution (optional).}

\item{lambda}{The shape parameter of the distribution (optional).}

\item{...}{Additional arguments to be passed to the limits function.}
}
\value{
A list containing the estimated parameters, alpha-hat, ARL, MRL, SDRL, control limits, convergence status, and value of the objective function.
}
\description{
This function calculates various statistical measures for the inverse Gaussian control chart.
}
\examples{
set.seed(0)
x <- r_ig(lots = 1000, n = 10, mu = 1, lambda = 1.7)

# Estimate for maximum likelihood
stats_ig(data = x)

# Using the true parameters
stats_ig(data = x, mu = 1, lambda = 1.7)

}
