% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{stats_n}
\alias{stats_n}
\title{Calculate statistics for a normal distribution control chart}
\usage{
stats_n(data, alpha = 0.0027, mu = NULL, sd = NULL, ...)
}
\arguments{
\item{data}{The data matrix}

\item{alpha}{The significance level}

\item{mu}{The mean of the distribution (optional)}

\item{sd}{The standard deviation of the distribution (optional)}

\item{...}{Additional arguments to be passed to the optimization algorithm}
}
\value{
A list containing the estimated parameters, control limits, alpha_hat, ARL, MRL, SDRL, convergence status, and value of the objective function
}
\description{
Calculate statistics for a normal distribution control chart
}
\examples{
set.seed(0)
x <- r_n(lots = 1000, n = 10, mu = 1, sd = 1.7)

# Estimate for maximum likelihood
stats_n(data = x, alpha = 0.0027)

# Useing the true parameters
stats_n(data = x, alpha = 0.0027, mu = 1, sd = 1.7)

}
